refactor(symbols): split symbol_analysis into mod.rs + tests.rs for LOC gate
PR #427 surfaced the workspace's "Reject .rs files over 1000 LOC" gate
because the new map-derived synthesis code + its 13 unit tests pushed
crates/fbuild-core/src/symbol_analysis.rs to 1076 lines (limit: 1000).

Promote `symbol_analysis` from a single file to a directory module:

- crates/fbuild-core/src/symbol_analysis/mod.rs  (695 LOC)
- crates/fbuild-core/src/symbol_analysis/tests.rs  (383 LOC)

`tests.rs` carries the existing `mod tests` body verbatim — no test
content changed; the `use super::*;` makes private items like
`nm_range_covers` and `InputSectionIndex` still visible inside the
child module. 22/22 tests still pass.

Pure refactor; no behavior change to the public API.
